joke_bot/requirements.txt
pyTeleBot==0.0.5
python-dotenv==1.0.0